pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markitdown-altpack"
dynamic = ["version"]
description = "Custom converters for the “markitdown” library, created by @arpanbiswas97."
authors = [
    {name = "Arpan Biswas",email = "arpanbiswas97@gmail.com"}
]
license = "MIT"
keywords = ["markitdown", "markitdown-plugin"]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "markitdown (>=0.1.0a1)",
    "lxml (>=5.3.1,<6.0.0)",
    "azure-ai-documentintelligence (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "pymupdf4llm (>=0.0.17,<0.0.18)",
    "azure-identity (>=1.21.0,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.1,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
]

[project.urls]
Documentation = "https://github.com/arpanbiswas97/markitdown-altpack#readme"
Issues = "https://github.com/arpanbiswas97/markitdown-altpack/issues"
Source = "https://github.com/arpanbiswas97/markitdown-altpack"

[tool.hatch.version]
path = "src/markitdown_altpack/__about__.py"

# IMPORTANT: MarkItDown will look for this entry point to find the plugin.
[project.entry-points."markitdown.plugin"]
markitdown_altpack = "markitdown_altpack"

[tool.hatch.envs.types]
extra-dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/markitdown_altpack tests}"

[tool.coverage.run]
source_pkgs = ["markitdown-altpack", "tests"]
branch = true
parallel = true
omit = ["src/markitdown_altpack/__about__.py"]

[tool.coverage.paths]
markitdown-sample-plugin = [
    "src/markitdown_altpack",
    "*/markitdown-altpack/src/markitdown_altpack",
]
tests = ["tests", "*/markitdown-altpack/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.hatch.build.targets.sdist]
only-include = ["src/markitdown_altpack"]

[tool.poetry]
version = "0.1.0a1"
